Moving _pad to utils.array.
